Fix codestyle of coll_cache_name

+++ src/box/coll_cache.h
@@ -74,7 +74,7 @@ coll_by_id(uint32_t id);
  * Find a collation object by its name.
  */
 struct coll *
-coll_by_name(const char *name, size_t len);
+coll_by_name(const char *name, uint32_t len);
 
 #if defined(__cplusplus)
 } /* extern "C" */
